Move to estimateReward instead pendingReward
estimateReward is a function of the Vertical Garden. It include the value of the reward debt of the master chef contract in the math.

So this function give a estimate of the gardener reward, without having to update the garden to update the pendingRewardToSplit value.

However it's a estimate value, it's a bit less precise.